+++ wagtail/api/v1/signal_handlers.py
