+++ wagtail/api/v1/tests/test_pages.py
@@ -8,7 +8,7 @@ from django.core.urlresolvers import reverse
 
 from wagtail.wagtailcore.models import Page
 
-from wagtail.contrib.wagtailapi import signal_handlers
+from wagtail.api.v1 import signal_handlers
 
 from wagtail.tests.demosite import models
 from wagtail.tests.testapp.models import StreamPage
